+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Initial NISAR converter based upon first 4 ASF sample datasets
+
 ### Changed
 - Modify the cosmo converter to accommodate some off nominal metadata.